rename incorrectly named index file
